# only use Table once