--- a/timvt/factory.py
+++ b/timvt/factory.py
@@ -216,7 +216,7 @@ class VectorTilerFactory:
                     return None
 
             return [
-                Table(**r.dict(by_alias=True), tileurl=_get_tiles_url(r.id))
+                Table(**r, tileurl=_get_tiles_url(r["id"]))
                 for r in request.app.state.table_catalog
             ]