This PR change the way we create the `Table` object. in #83 we were first creating a `dbmodel.Table` and then passing it (using `table.dict()`) to `layer.Table`. This result in useless double validation (slow down application)


To be discussed

cc @bitner