@vincentsarago 2022-06-15
## Master

```
Run siege --file .github/workflows/data/urls_epsg4326.txt -b -c 20 -r 100 > /dev/null
** SIEGE 4.1.3
** Preparing 20 concurrent users for battle.
The server is now under siege...

Transactions:                   2000 hits
Availability:                 100.00 %
Elapsed time:                   2.81 secs
Data transferred:               5.47 MB
Response time:                  0.02 secs
Transaction rate:             711.74 trans/sec
Throughput:                     1.95 MB/sec
Concurrency:                   17.10
Successful transactions:        2000
Failed transactions:               0
Longest transaction:            0.12
Shortest transaction:           0.01
```

## #83
```
Run siege --file .github/workflows/data/urls_epsg4326.txt -b -c 20 -r 100 > /dev/null
** SIEGE 4.1.3
** Preparing 20 concurrent users for battle.
The server is now under siege...

Transactions:                   2000 hits
Availability:                 100.00 %
Elapsed time:                   3.53 secs
Data transferred:               5.47 MB
Response time:                  0.03 secs
Transaction rate:             566.57 trans/sec
Throughput:                     1.55 MB/sec
Concurrency:                   18.78
Successful transactions:        2000
Failed transactions:               0
Longest transaction:            0.09
Shortest transaction:           0.01
```

## This PR 
```
Run siege --file .github/workflows/data/urls_epsg4326.txt -b -c 20 -r 100 > /dev/null
** SIEGE 4.1.3
** Preparing 20 concurrent users for battle.
The server is now under siege...

Transactions:                   2000 hits
Availability:                 100.00 %
Elapsed time:                   3.14 secs
Data transferred:               5.47 MB
Response time:                  0.03 secs
Transaction rate:             636.94 trans/sec
Throughput:                     1.74 MB/sec
Concurrency:                   18.12
Successful transactions:        2000
Failed transactions:               0
Longest transaction:            0.19
Shortest transaction:           0.01
```
☝️ shows benchmark ran locally on my Mac OS. Benchmark in CI is much slower and doesn't seems to show huge difference between branches


This branch (`reduceModelUsage`) seems to be faster than #83 but still slower than master 🤷‍♂️