@vincentsarago 2022-06-15
@bitner I know this is the opposite of what you wanted first but I think having to pay twice for model validation (assuming this is the cause of the slow down) is too much. With this PR we now store the `Database` tables as Dict and only do the `validation` once per table within specific request (Layer dependency).